Block third-level domains without leading period.
Most third-level ad-serving domains include an additional fourth, and
occasionally fifth-level name, and the leading `.` character reliably
matched all of the variations.

The ads.yahoo.com third-level domain, however, serves ads directly and
is not matched due to the expected leading `.` character. So, include a
leading period in the domain expression only for second-level domains.