Improve domain match patterns.
Match literal dot in domain names by escaping it. Anchor domain to
front of URL so it won’t match in a query string.

We previously processed a list of raw domain names into this form, but
these constraints must be included in each url-filter pattern for the
Content Blocking API.

/cc @BenjaminPoulain